Added layout component
Took 2 hours 51 minutes
