--- Tic-Tac-Toe-React/dev-tic-tac-toe/public/robots.txt
@@ -1,3 +1,0 @@
-# https://www.robotstxt.org/robotstxt.html
-User-agent: *
-Disallow:
