--- Tic-Tac-Toe-React/dev-tic-tac-toe/src/Component.js
@@ -1,7 +1,0 @@
-export default function Component() {
-    return (
-        <div>
-            <h1> This is a component</h1>
-        </div>
-    )
-};
