--- Tic-Tac-Toe-React/dev-tic-tac-toe/src/NavBar.js
@@ -1,7 +1,0 @@
-export default function NavBar() {
-    return (
-        <div>
-            <h1> This is a component</h1>
-        </div>
-    )
-};
