+++ Tic-Tac-Toe/dev-tic-tac-toe/public/robots.txt
@@ -1,0 +1,3 @@
+# https://www.robotstxt.org/robotstxt.html
+User-agent: *
+Disallow:
