+++ Tic-Tac-Toe/dev-tic-tac-toe/src/Component.js
@@ -1,0 +1,7 @@
+export default function Component() {
+    return (
+        <div>
+            <h1> This is a component</h1>
+        </div>
+    )
+};
