+++ Tic-Tac-Toe/dev-tic-tac-toe/src/NavBar.js
@@ -1,0 +1,7 @@
+export default function NavBar() {
+    return (
+        <div>
+            <h1> This is a component</h1>
+        </div>
+    )
+};
