+++ Tic-Tac-Toe/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>Tic-Tac-Toe</title>
     <link href="style.css" rel="stylesheet">
-    <!--    <script src="script.js"></script>-->
+    <!--    <sc11ript 10src="script.js"></script>-->
     <!--    <script src="scriptTwo.js"></script>-->
     <!--     <script src="scriptThree.js" ></script>-->
     <!--    <script src="scriptFour.js" defer></script>-->
@@ -17,8 +17,9 @@
 <html lang="en">
 <body>
 <nav>
+    <img id="falcon-two" src="images/DT%20Logo%206.png" alt="Logo">
+    <h3 id="title">UWRF Development Team</h3>
     <ul>
-        <li class="nav-items"><img id="falcon-two" src="images/DT%20Logo%205.png" alt="Logo"></li>
         <li class="nav-items"><h3> Home </h3> </li>
         <li class="nav-items"><h3> Projects</h3></li>
     </ul>
@@ -26,8 +27,8 @@
 <!--<img id="falcon" alt="falcon-logo" src="DT%20Logo%205.png">-->
 
 <header>
-    <h1>UWRF Development Team</h1>
-    <h3>Tic-Tac-Toe</h3>
+<!--    <h1>UWRF Development Team</h1>-->
+    <h1 id="tic-tac-toe-title">Tic-Tac-Toe</h1>
 </header>
 
 <div id="body-top">
